JSON Output
Reads a CSV with at least "Address" and "City" fields and converts it to JSON format. To include lat/longs, the "-g" or "--geocode" parameter needs to be set to "on." Default is "off" since it requires a webservice call to ArcGIS; it's faster for testing purpose to have default off.